+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## [Unreleased] - ReleaseDate
+
+### Changed
+
+* `spanr merge`: take the first part of filename
+* `spanr compare`: more than two infiles
+
 ## [0.3.2] - 2019-09-03
 
 ### Added